chore: relocate internal/testutil to test/testutil
