+++ internal/collector/source/pod/pod_scraping_source_test.go
@@ -19,7 +19,7 @@ import (
 	sourcepkg "github.com/llm-d/llm-d-workload-variant-autoscaler/internal/collector/source"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/constants"
 	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/metrics"
-	"github.com/llm-d/llm-d-workload-variant-autoscaler/internal/testutil"
+	"github.com/llm-d/llm-d-workload-variant-autoscaler/test/testutil"
 )
 
 var _ = Describe("PodScrapingSource", func() {
